Refactor for new products - frontend done
