+++ glam/api/views.py
@@ -5,26 +5,21 @@ import dateutil.parser
 import orjson
 from django.conf import settings
 from django.core.cache import caches
-from django.db.models import Max, Q, Value, Count
-from django.db.models.functions import Concat
+from django.db.models import Max, Q, Count
 from django.views.decorators.cache import cache_page
 from rest_framework.decorators import api_view
 from rest_framework.exceptions import NotFound, ValidationError
 from rest_framework.response import Response
 from google.cloud import bigquery
+from glam.api.bigquery import build_glean_query, execute_glean_query
 
 
 from glam.api import constants
 from glam.api.models import (
-    DesktopBetaAggregationView,
     DesktopNightlyAggregationView,
-    DesktopReleaseAggregationView,
     FenixAggregationView,
     FOGAggregationView,
-    FenixCounts,
-    FOGCounts,
     FirefoxBuildRevisions,
-    FirefoxCounts,
     LastUpdated,
     Probe,
     UsageInstrumentation,
@@ -35,7 +30,7 @@ GLAM_BQ_PROD_PROJECT = "moz-fx-data-shared-prod"
 
 
 def get_bq_client():
-    return bigquery.Client()
+    return bigquery.Client("efilho-sandbox")
 
 
 @api_view(["GET"])
@@ -146,120 +141,13 @@ def validate_request_glean(**kwargs):
     return validated_data
 
 
-def get_firefox_aggregations(source, request, **kwargs):
-    if source == "BigQuery":
-        bqClient = get_bq_client()
-        req_data = validate_request_legacy(**kwargs)
-        return get_firefox_aggregations_from_bq(bqClient, request, req_data)
-    else:
-        return get_firefox_aggregations_from_pg(request, **kwargs)
+def get_firefox_aggregations(request, **kwargs):
+    req_data = validate_request_legacy(**kwargs)
+    return get_firefox_aggregations_from_bq(request, req_data)
 
 
-def get_firefox_aggregations_from_pg(request, **kwargs):
-    # TODO: When glam starts sending "product", make it required.
-    REQUIRED_QUERY_PARAMETERS = ["channel", "probe", "aggregationLevel"]
-    if any([k not in kwargs.keys() for k in REQUIRED_QUERY_PARAMETERS]):
-        # Figure out which query parameter is missing.
-        missing = set(REQUIRED_QUERY_PARAMETERS) - set(kwargs.keys())
-        raise ValidationError(
-            "Missing required query parameters: {}".format(", ".join(sorted(missing)))
-        )
-
-    # Ensure that the product provided is one we support, defaulting to Firefox.
-    product = "firefox"
-    channel = kwargs.get("channel")
-    model_key = f"{product}-{channel}"
-
-    MODEL_MAP = {
-        "firefox-nightly": DesktopNightlyAggregationView,
-        "firefox-beta": DesktopBetaAggregationView,
-        "firefox-release": DesktopReleaseAggregationView,
-    }
-
-    try:
-        model = MODEL_MAP[model_key]
-    except KeyError:
-        raise ValidationError("Product not currently supported.")
-
-    num_versions = kwargs.get("versions", 3)
-    try:
-        max_version = int(model.objects.aggregate(Max("version"))["version__max"])
-    except (ValueError, KeyError):
-        raise ValidationError("Query version cannot be determined")
-    except TypeError:
-        # This happens when `version_max` is NULL and cannot be converted to an int,
-        # suggesting that we have no data for this model.
-        raise NotFound("No data found for the provided parameters")
-
-    versions = list(map(str, range(max_version, max_version - num_versions, -1)))
-
-    labels_cache = caches["probe-labels"]
-    if labels_cache.get("__labels__") is None:
-        Probe.populate_labels_cache()
-
-    os = kwargs.get("os", "*")
-
-    dimensions = [
-        Q(metric=kwargs["probe"]),
-        Q(version__in=versions),
-        Q(os=os),
-    ]
-
-    aggregation_level = kwargs["aggregationLevel"]
-    # Whether to pull aggregations by version or build_id.
-    if aggregation_level == "version":
-        dimensions.append(Q(build_id="*"))
-        # counts = _get_firefox_counts(channel, os, versions, by_build=False)
-        shas = {}
-    elif aggregation_level == "build_id":
-        dimensions.append(~Q(build_id="*"))
-        # counts = _get_firefox_counts(channel, os, versions, by_build=True)
-        shas = _get_firefox_shas(channel)
-
-    if "process" in kwargs:
-        dimensions.append(Q(process=kwargs["process"]))
-    result = model.objects.filter(*dimensions)
-
-    response = []
-
-    for row in result:
-
-        data = {
-            "version": row.version,
-            "os": row.os,
-            "build_id": row.build_id,
-            "revision": shas.get(row.build_id, ""),
-            "process": row.process,
-            "metric": row.metric,
-            "metric_key": row.metric_key,
-            "metric_type": row.metric_type,
-            "total_users": row.total_users,
-            "sample_count": row.total_sample,
-            "histogram": row.histogram and orjson.loads(row.histogram) or "",
-            "non_norm_histogram": row.non_norm_histogram
-            and orjson.loads(row.non_norm_histogram)
-            or "",
-            "percentiles": row.percentiles and orjson.loads(row.percentiles) or "",
-            "non_norm_percentiles": row.non_norm_percentiles
-            and orjson.loads(row.non_norm_percentiles)
-            or "",
-        }
-        if row.client_agg_type:
-            if row.metric_type == "boolean":
-                data["client_agg_type"] = "boolean-histogram"
-            else:
-                data["client_agg_type"] = row.client_agg_type
-
-        # Get the total distinct client IDs for this set of dimensions.
-        # data["total_addressable_market"] = counts.get(f"{row.version}-{row.build_id}")
-
-        response.append(data)
-
-    _log_probe_query(request)
-    return response
-
-
-def get_firefox_aggregations_from_bq(bqClient, request, req_data):
+def get_firefox_aggregations_from_bq(request, req_data):
+    bqClient = get_bq_client()
     channel = req_data["channel"]
     aggregation_level = req_data["aggregation_level"]
     # Whether to pull aggregations by version or build_id.
@@ -367,29 +255,6 @@ def get_firefox_aggregations_from_bq(bqClient, request, req_data):
     return response
 
 
-def _get_firefox_counts(channel, os, versions, by_build):
-    """
-    Helper method to gather the `FirefoxCounts` data in a single query.
-
-    Returns the data as a Python dict keyed by "{version}-{build_id}" for
-    quick lookup.
-
-    """
-    query = FirefoxCounts.objects.filter(
-        channel=constants.CHANNEL_IDS[channel], os=os, version__in=versions
-    )
-    if by_build:
-        query = query.exclude(build_id="*")
-    else:
-        query = query.filter(build_id="*")
-    query = query.annotate(key=Concat("version", Value("-"), "build_id"))
-    data = {
-        row["key"]: row["total_users"] for row in query.values("key", "total_users")
-    }
-
-    return data
-
-
 def _get_firefox_shas(channel, hourly=False):
     shas_full = dict(
         FirefoxBuildRevisions.objects.filter(channel=channel).values_list(
@@ -402,116 +267,13 @@ def _get_firefox_shas(channel, hourly=False):
         return {build[:10]: revision for build, revision in shas_full.items()}
 
 
-def get_glean_aggregations(source, request, **kwargs):
-    if source == "BigQuery":
-        bqClient = get_bq_client()
-        req_data = validate_request_glean(**kwargs)
-        return get_glean_aggregations_from_bq(bqClient, request, req_data)
-    else:
-        return get_glean_aggregations_from_pg(request, **kwargs)
+def get_glean_aggregations(request, **kwargs):
+    req_data = validate_request_glean(**kwargs)
+    return get_glean_aggregations_from_bq(request, req_data)
 
 
-def get_glean_aggregations_from_pg(request, **kwargs):
-    REQUIRED_QUERY_PARAMETERS = [
-        "aggregationLevel",
-        "app_id",
-        "ping_type",
-        "probe",
-        "product",
-    ]
-    if any([k not in kwargs.keys() for k in REQUIRED_QUERY_PARAMETERS]):
-        # Figure out which query parameter is missing.
-        missing = set(REQUIRED_QUERY_PARAMETERS) - set(kwargs.keys())
-        raise ValidationError(
-            "Missing required query parameters: {}".format(", ".join(sorted(missing)))
-        )
-
-    MODEL_MAP = {
-        "fenix": FenixAggregationView,
-        "fog": FOGAggregationView,
-    }
-    model = MODEL_MAP[kwargs.get("product")]
-    product = kwargs.get("product")
-    probe = kwargs["probe"]
-
-    num_versions = kwargs.get("versions", 3)
-    try:
-        versions = list(
-            model.objects.filter(Q(metric=probe))
-            .order_by("-version")
-            .values_list("version", flat=True)
-            .distinct("version")[:num_versions]
-        )
-    except (ValueError, KeyError):
-        raise ValidationError("Query version cannot be determined")
-    except TypeError:
-        # This happens when `version` is NULL,
-        # suggesting that we have no data for this model.
-        raise NotFound("No data found for the provided parameters")
-
-    app_id = kwargs["app_id"]
-    ping_type = kwargs["ping_type"]
-    os = kwargs.get("os", "*")
-
-    dimensions = [
-        Q(app_id=app_id),
-        Q(metric=probe),
-        Q(ping_type=ping_type),
-        Q(version__in=versions),
-        Q(os=os),
-    ]
-
-    aggregation_level = kwargs["aggregationLevel"]
-    # Whether to pull aggregations by version or build_id.
-    if aggregation_level == "version":
-        if product == "fenix":
-            dimensions.append(Q(build_id="*"))
-            # counts = _get_fenix_counts(app_id, versions, ping_type,
-            #  os, by_build=False)
-        if product == "fog":
-            dimensions.append(~Q(build_id="*"))
-            # counts = _get_fog_counts(app_id, versions, ping_type, os, by_build=False)
-
-    if aggregation_level == "build_id":
-        if product == "fenix":
-            dimensions.append(~Q(build_id="*"))
-            # counts = _get_fenix_counts(app_id, versions, ping_type, os, by_build=True)
-        if product == "fog":
-            dimensions.append(~Q(build_id="*"))
-            # counts = _get_fog_counts(app_id, versions, ping_type, os, by_build=True)
-
-    result = model.objects.filter(*dimensions)
-
-    response = []
-
-    for row in result:
-
-        data = {
-            "version": row.version,
-            "ping_type": row.ping_type,
-            "os": row.os,
-            "build_id": row.build_id,
-            "build_date": row.build_date,
-            "metric": row.metric,
-            "metric_type": row.metric_type,
-            "metric_key": row.metric_key,
-            "client_agg_type": row.client_agg_type,
-            "total_users": row.total_users,
-            "sample_count": row.total_sample,
-            "histogram": row.histogram and orjson.loads(row.histogram) or "",
-            "percentiles": row.percentiles and orjson.loads(row.percentiles) or "",
-        }
-
-        # Get the total distinct client IDs for this set of dimensions.
-        # data["total_addressable_market"] = counts.get(f"{row.version}-{row.build_id}")
-
-        response.append(data)
-    _log_probe_query(request)
-    return response
-
-
-def get_glean_aggregations_from_bq(bqClient, request, req_data):
-
+def get_glean_aggregations_from_bq(request, req_data):
+    bqClient = get_bq_client()
     channel = req_data["channel"]
     product = req_data["product"]
     probe = req_data["probe"]
@@ -620,52 +382,6 @@ def get_glean_aggregations_from_bq(bqClient, request, req_data):
     return response
 
 
-def _get_fenix_counts(app_id, versions, ping_type, os, by_build):
-    """
-    Helper method to gather the `FenixCounts` data in a single query.
-
-    Returns the data as a Python dict keyed by "{version}-{build_id}" for
-    quick lookup.
-
-    """
-    query = FenixCounts.objects.filter(
-        app_id=app_id, version__in=versions, ping_type=ping_type, os=os
-    )
-    if by_build:
-        query = query.exclude(build_id="*")
-    else:
-        query = query.filter(build_id="*")
-    query = query.annotate(key=Concat("version", Value("-"), "build_id"))
-    data = {
-        row["key"]: row["total_users"] for row in query.values("key", "total_users")
-    }
-
-    return data
-
-
-def _get_fog_counts(app_id, versions, ping_type, os, by_build):
-    """
-    Helper method to gather the `FOGCounts` data in a single query.
-
-    Returns the data as a Python dict keyed by "{version}-{build_id}" for
-    quick lookup.
-
-    """
-    query = FOGCounts.objects.filter(
-        app_id=app_id, version__in=versions, ping_type=ping_type, os=os
-    )
-    if by_build:
-        query = query.exclude(build_id="*")
-    else:
-        query = query.filter(build_id="*")
-    query = query.annotate(key=Concat("version", Value("-"), "build_id"))
-    data = {
-        row["key"]: row["total_users"] for row in query.values("key", "total_users")
-    }
-
-    return data
-
-
 @api_view(["POST"])
 def aggregations(request):
     """
@@ -736,13 +452,12 @@ def aggregations(request):
             )
         )
 
-    data_source = body.get("data_source", "BigQuery")
     # Uncomment the next line to rollback to PostgreSQL
     # data_source = "Postgres"
     if product == FIREFOX_LEGACY:
-        response = get_firefox_aggregations(data_source, request, **body["query"])
+        response = get_firefox_aggregations(request, **body["query"])
     else:  # Assume everything else is Glean-based.
-        response = get_glean_aggregations(data_source, request, **body["query"])
+        response = get_glean_aggregations(request, **body["query"])
 
     if not response:
         raise NotFound("No documents found for the given parameters")
@@ -775,7 +490,7 @@ def _get_random_probes(data_source, random_percentage, limit):
         table_name = (
             f"`{GLAM_BQ_PROD_PROJECT}.glam_etl.glam_desktop_nightly_aggregates`"
         )
-        with bigquery.Client() as client:
+        with get_bq_client() as client:
             aggs = client.query(
                 query_template.format(
                     "", table_name, "RAND()", random_percentage, limit
@@ -864,7 +579,7 @@ def _get_fx_most_used_probes(days=30, limit=9):
         ]
     )
 
-    with bigquery.Client() as client:
+    with get_bq_client() as client:
         aggs = client.query(query, job_config=job_config)
     return aggs
 
@@ -944,3 +659,74 @@ def usage(request):
         else:
             response = result.values("action_type", "timestamp", "probe_name")
         return Response(response, 200)
+
+
+def validate_glean_request(request):
+    """Validate request parameters for Glean-based products."""
+    required_params = {
+        'aggregationLevel': str,
+        'app_id': str,
+        'ping_type': str,
+        'metric': str,
+        'metric_type': str,
+        'os': str,
+        'timeHorizon': str,
+    }
+
+    for param, param_type in required_params.items():
+        if param not in request:
+            raise ValueError(f"Missing required parameter: {param}")
+        if not isinstance(request[param], param_type):
+            raise ValueError(f"Invalid type for {param}: expected {param_type}")
+
+    # Validate product-specific parameters
+    product = request.get('product')
+    if product not in ['fog', 'fenix']:
+        raise ValueError(f"Invalid product: {product}")
+
+    # Validate app_id based on product
+    valid_app_ids = {
+        'fog': ['nightly', 'beta', 'release'],
+        'fenix': ['nightly', 'beta', 'release']
+    }
+    if request['app_id'] not in valid_app_ids[product]:
+        raise ValueError(f"Invalid app_id for {product}: {request['app_id']}")
+
+    # Validate ping_type based on product
+    valid_ping_types = {
+        'fog': ['baseline', 'metrics'],
+        'fenix': ['baseline', 'metrics', 'events']
+    }
+    if request['ping_type'] not in valid_ping_types[product]:
+        raise ValueError(f"Invalid ping_type for {product}: {request['ping_type']}")
+
+    return True
+
+def get_glean_data(request):
+    """Get data for Glean-based products."""
+    validate_glean_request(request)
+
+    product = request['product']
+    metric = request['metric']
+    metric_type = request['metric_type']
+    app_id = request['app_id']
+    ping_type = request['ping_type']
+    os = request['os']
+    time_horizon = request['timeHorizon']
+
+    # Get product-specific configuration
+    product_config = get_product_config(product)
+
+    # Build query based on product configuration
+    query = build_glean_query(
+        product_config,
+        metric,
+        metric_type,
+        app_id,
+        ping_type,
+        os,
+        time_horizon
+    )
+
+    # Execute query and return results
+    return execute_glean_query(query)
