--- src/config/glean.js
@@ -1,494 +1,0 @@
-import produce from 'immer';
-import {
-  transformAPIResponse,
-  transformLabeledCounterToCategoricalHistogram,
-} from '../utils/transform-data';
-import { stripDefaultValues } from '../utils/urls';
-import sharedDefaults, { extractBucketMetadata } from './shared';
-import { getProbeData, getProbeInfo } from '../state/api';
-
-import {
-  validate,
-  noResponse,
-  noUnknownMetrics,
-} from '../utils/data-validation';
-
-export const SUPPORTED_METRICS = [
-  'categorical',
-  'linear',
-  'counter',
-  'labeled_counter',
-  'quantity',
-  'timespan',
-  'custom_distribution',
-  'memory_distribution',
-  'timing_distribution',
-  'labeled_custom_distribution',
-  'labeled_timing_distribution',
-];
-
-export const FIREFOX_ON_GLEAN = {
-  label: 'Firefox on Glean',
-  key: 'fog',
-  sampleRate: 1,
-  dimensions: {
-    normalizationType: {
-      title: 'Normalization',
-      key: 'normalizationType',
-      values: [
-        { key: 'normalized', label: 'By Client ID' },
-        { key: 'non_normalized', label: 'None' },
-      ],
-      defaultValue: 'normalized',
-    },
-    app_id: {
-      title: 'App',
-      key: 'app_id',
-      values: [
-        { key: 'nightly', label: 'Nightly' },
-        { key: 'beta', label: 'Beta' },
-        { key: 'release', label: 'Release' },
-      ],
-      defaultValue: 'nightly',
-    },
-    os: {
-      title: 'OS',
-      key: 'os',
-      values: [
-        { key: '*', label: 'All OSes' },
-        { key: 'Windows', label: 'Windows' },
-        { key: 'Darwin', label: 'Mac' },
-        { key: 'Linux', label: 'Linux' },
-      ],
-      defaultValue: '*',
-    },
-    ping_type: {
-      title: 'Ping Type',
-      key: 'ping_type',
-      values: [
-        { key: '*', label: 'All' },
-        { key: 'metrics', label: 'Metrics' },
-        { key: 'baseline', label: 'Baseline' },
-      ],
-      defaultValue: 'metrics',
-      isValidKey(key, probe) {
-        return key === '*' ? true : probe.send_in_pings.includes(key);
-      },
-    },
-    aggregationLevel: {
-      title: 'Aggregation Level',
-      key: 'aggregationLevel',
-      values: [
-        { key: 'build_id', label: 'Build ID' },
-        { key: 'version', label: 'Major Version' },
-      ],
-      defaultValue: 'build_id',
-    },
-  },
-  // FIXME: these are guesses at the moment
-  probeViewMap: {
-    boolean: 'categorical',
-    counter: 'linear',
-    custom_distribution_exponential: 'log',
-    custom_distribution_linear: 'linear',
-    labeled_counter: (isStatic) => (isStatic ? 'categorical' : 'linear'),
-    memory_distribution: 'log',
-    quantity: 'linear',
-    timespan: 'log',
-    timing_distribution: 'log',
-    labeled_timing_distribution: 'log',
-  },
-  probeViewFromHistogramTypeMap: {
-    exponential: 'log',
-    linear: 'linear',
-  },
-
-  getViewFromMetricType(metricType, isStatic = false) {
-    const mode = this.probeViewMap[metricType];
-    return typeof mode === 'function' ? mode(isStatic) : mode;
-  },
-
-  getParamsForQueryString(storeValue) {
-    // These parameters will map to a ${key}=${value}&... in the querystring,
-    // which is used to convey the view state when the GLAM URL is shared with
-    // others.
-    const params = {
-      app_id: storeValue.productDimensions.app_id,
-      os: storeValue.productDimensions.os,
-      ping_type: storeValue.productDimensions.ping_type,
-      aggregationLevel: storeValue.productDimensions.aggregationLevel,
-      timeHorizon: storeValue.timeHorizon,
-      proportionMetricType: storeValue.proportionMetricType,
-      activeBuckets: storeValue.activeBuckets,
-      visiblePercentiles: storeValue.visiblePercentiles,
-      ref: storeValue.ref,
-      aggKey: storeValue.aggKey,
-      aggType: storeValue.aggType,
-      currentPage: storeValue.currentPage,
-      normalizationType: storeValue.productDimensions.normalizationType,
-    };
-    return stripDefaultValues(params, {
-      ...sharedDefaults,
-      ...this.dimensions,
-    });
-  },
-  getParamsForDataAPI(storeValue) {
-    // These parameters are needed to request the data from the API itself
-    return {
-      product: storeValue.product,
-      app_id: storeValue.productDimensions.app_id,
-      os: storeValue.productDimensions.os,
-      ping_type: storeValue.productDimensions.ping_type,
-      probe: storeValue.probeName,
-      aggregationLevel: storeValue.productDimensions.aggregationLevel,
-      versions: 20,
-    };
-  },
-  async fetchData(params, appStore) {
-    await getProbeInfo(appStore.getState().product, params.probe).then((r) => {
-      const newProbe = { ...r, loaded: true };
-      appStore.setField('probe', newProbe);
-    });
-
-    const metricType = appStore.getState().probe.type;
-    const histogramType = appStore.getState().probe.histogram_type;
-    let probeView = this.getViewFromMetricType(metricType)
-      ? this.getViewFromMetricType(metricType)
-      : this.probeViewFromHistogramTypeMap[histogramType];
-    if (metricType === 'labeled_counter') {
-      const isStatic =
-        appStore.getState().probe.labels !== null &&
-        appStore.getState().probe.labels.length > 0;
-      probeView = this.getViewFromMetricType(metricType, isStatic);
-    }
-    noUnknownMetrics(metricType, SUPPORTED_METRICS);
-
-    return getProbeData(params).then((payload) => {
-      const { aggregationLevel } = appStore.getState().productDimensions;
-
-      validate(payload, (p) => {
-        noResponse(p);
-        noUnknownMetrics(metricType, SUPPORTED_METRICS);
-      });
-      const viewType = probeView === 'categorical' ? 'proportion' : 'quantile';
-      let data = payload.response;
-
-      appStore.setField('viewType', viewType);
-      appStore.setField('aggMethod', data[0].client_agg_type);
-
-      // Attach labels to histogram if appropriate type.
-      if (probeView === 'categorical') {
-        const labels = {
-          ...appStore.getState().probe.labels,
-        };
-        if (metricType === 'labeled_counter') {
-          data = transformLabeledCounterToCategoricalHistogram(data, labels);
-        }
-        data = produce(data, (draft) =>
-          draft.map((point) => ({
-            ...point,
-            histogram: Object.entries(point.histogram).reduce(
-              (acc, [bin, value]) => {
-                const intBin = Math.floor(bin);
-                if (intBin in labels) {
-                  acc[labels[intBin]] = value;
-                }
-                return acc;
-              },
-              {}
-            ),
-            non_norm_histogram: Object.entries(point.non_norm_histogram).reduce(
-              (acc, [bin, value]) => {
-                const intBin = Math.floor(bin);
-                if (intBin in labels) {
-                  acc[labels[intBin]] = value;
-                }
-                return acc;
-              },
-              {}
-            ),
-          }))
-        );
-      }
-      data = transformAPIResponse[viewType](data, aggregationLevel, metricType);
-      return {
-        data,
-        probeType: probeView,
-      };
-    });
-  },
-
-  updateStoreAfterDataIsReceived(data, probeType, appStore) {
-    // This function is called directly after the response has been received by
-    // the frontend. It will always run, even against cached data, as a way of
-    // resetting the necessary state.
-    const viewType = probeType;
-    let etc = {};
-
-    // filter out true/false aggregate results in boolean metrics. See: https://github.com/mozilla/glam/pull/1525#discussion_r694135079
-    if (data[0].metric_type === 'boolean') {
-      // eslint-disable-next-line no-param-reassign
-      data = data.filter((di) => di.client_agg_type === '');
-    }
-
-    if (viewType === 'categorical') {
-      etc = extractBucketMetadata(data);
-      appStore.setField('activeBuckets', etc.initialBuckets);
-    }
-
-    return { data, viewType, ...etc };
-  },
-  transformProbeForGLAM(probe) {
-    const pr = { ...probe };
-    return pr;
-  },
-  setDefaultsForProbe(store) {
-    const state = store.getState();
-    const { probe } = state;
-    if (!probe.send_in_pings.includes(state.productDimensions.ping_type)) {
-      // Try not to pick the 'All' option, at first.
-      const index = probe.send_in_pings.length > 1 ? 1 : 0;
-      let newPingtype = probe.send_in_pings[index];
-      if (
-        !Object.keys(this.dimensions.ping_type.values).includes(newPingtype)
-      ) {
-        // We don't know about this ping type, default to 'All'.
-        newPingtype = '*';
-      }
-      store.setDimension('ping_type', newPingtype);
-    }
-  },
-};
-
-export const FENIX = {
-  label: 'Firefox Android',
-  key: 'fenix',
-  sampleRate: 1,
-  dimensions: {
-    normalizationType: {
-      title: 'Normalization',
-      key: 'normalizationType',
-      values: [
-        { key: 'normalized', label: 'By Client ID' },
-        { key: 'non_normalized', label: 'None' },
-      ],
-      defaultValue: 'normalized',
-    },
-    app_id: {
-      title: 'App',
-      key: 'app_id',
-      values: [
-        { key: 'nightly', label: 'Nightly' },
-        { key: 'beta', label: 'Beta' },
-        { key: 'release', label: 'Release' },
-      ],
-      defaultValue: 'nightly',
-    },
-    os: {
-      title: 'OS',
-      key: 'os',
-      values: [{ key: 'Android', label: 'Android' }],
-      defaultValue: 'Android',
-    },
-    ping_type: {
-      title: 'Ping Type',
-      key: 'ping_type',
-      values: [
-        { key: '*', label: 'All' },
-        { key: 'baseline', label: 'Baseline' },
-        { key: 'metrics', label: 'Metrics' },
-        { key: 'first-session', label: 'First Session' },
-        { key: 'migration', label: 'Migration' },
-      ],
-      defaultValue: 'metrics',
-      isValidKey(key, probe) {
-        return key === '*' ? true : probe.send_in_pings.includes(key);
-      },
-    },
-    aggregationLevel: {
-      title: 'Aggregation Level',
-      key: 'aggregationLevel',
-      values: [
-        { key: 'build_id', label: 'Build ID' },
-        { key: 'version', label: 'Major Version' },
-      ],
-      defaultValue: 'build_id',
-    },
-  },
-  // FIXME: these are guesses at the moment
-  probeViewMap: {
-    boolean: 'categorical',
-    counter: 'linear',
-    custom_distribution_exponential: 'log',
-    custom_distribution_linear: 'linear',
-    labeled_counter: (isStatic) => (isStatic ? 'categorical' : 'linear'),
-    memory_distribution: 'log',
-    quantity: 'linear',
-    timespan: 'log',
-    timing_distribution: 'log',
-    labeled_timing_distribution: 'log',
-  },
-  probeViewFromHistogramTypeMap: {
-    exponential: 'log',
-    linear: 'linear',
-  },
-
-  getViewFromMetricType(metricType, isStatic = false) {
-    const mode = this.probeViewMap[metricType];
-    return typeof mode === 'function' ? mode(isStatic) : mode;
-  },
-
-  getParamsForQueryString(storeValue) {
-    // These parameters will map to a ${key}=${value}&... in the querystring,
-    // which is used to convey the view state when the GLAM URL is shared with
-    // others.
-    const params = {
-      app_id: storeValue.productDimensions.app_id,
-      os: storeValue.productDimensions.os,
-      ping_type: storeValue.productDimensions.ping_type,
-      aggregationLevel: storeValue.productDimensions.aggregationLevel,
-      timeHorizon: storeValue.timeHorizon,
-      proportionMetricType: storeValue.proportionMetricType,
-      activeBuckets: storeValue.activeBuckets,
-      visiblePercentiles: storeValue.visiblePercentiles,
-      ref: storeValue.ref,
-      aggKey: storeValue.aggKey,
-      aggType: storeValue.aggType,
-      currentPage: storeValue.currentPage,
-      normalizationType: storeValue.productDimensions.normalizationType,
-    };
-    return stripDefaultValues(params, {
-      ...sharedDefaults,
-      ...this.dimensions,
-    });
-  },
-  getParamsForDataAPI(storeValue) {
-    // These parameters are needed to request the data from the API itself
-    return {
-      product: storeValue.product,
-      app_id: storeValue.productDimensions.app_id,
-      os: storeValue.productDimensions.os,
-      ping_type: storeValue.productDimensions.ping_type,
-      probe: storeValue.probeName,
-      aggregationLevel: storeValue.productDimensions.aggregationLevel,
-      versions: 20,
-    };
-  },
-  async fetchData(params, appStore) {
-    await getProbeInfo(appStore.getState().product, params.probe).then((r) => {
-      const newProbe = { ...r, loaded: true };
-      appStore.setField('probe', newProbe);
-    });
-
-    const metricType = appStore.getState().probe.type;
-    const histogramType = appStore.getState().probe.histogram_type;
-    let probeView = this.getViewFromMetricType(metricType)
-      ? this.getViewFromMetricType(metricType)
-      : this.probeViewFromHistogramTypeMap[histogramType];
-    if (metricType === 'labeled_counter') {
-      const isStatic =
-        appStore.getState().probe.labels !== null &&
-        appStore.getState().probe.labels.length > 0;
-      probeView = this.getViewFromMetricType(metricType, isStatic);
-    }
-    noUnknownMetrics(metricType, SUPPORTED_METRICS);
-
-    return getProbeData(params).then((payload) => {
-      const { aggregationLevel } = appStore.getState().productDimensions;
-
-      validate(payload, (p) => {
-        noResponse(p);
-        noUnknownMetrics(metricType, SUPPORTED_METRICS);
-      });
-      const viewType = probeView === 'categorical' ? 'proportion' : 'quantile';
-      let data = payload.response;
-
-      appStore.setField('viewType', viewType);
-      appStore.setField('aggMethod', payload.response[0].client_agg_type);
-      // Attach labels to histogram if appropriate type.
-      if (probeView === 'categorical') {
-        const labels = {
-          ...appStore.getState().probe.labels,
-        };
-        if (metricType === 'labeled_counter') {
-          data = transformLabeledCounterToCategoricalHistogram(data, labels);
-        }
-        data = produce(data, (draft) =>
-          draft.map((point) => ({
-            ...point,
-            histogram: Object.entries(point.histogram).reduce(
-              (acc, [bin, value]) => {
-                const intBin = Math.floor(bin);
-                if (intBin in labels) {
-                  acc[labels[intBin]] = value;
-                }
-                return acc;
-              },
-              {}
-            ),
-            non_norm_histogram: Object.entries(point.non_norm_histogram).reduce(
-              (acc, [bin, value]) => {
-                const intBin = Math.floor(bin);
-                if (intBin in labels) {
-                  acc[labels[intBin]] = value;
-                }
-                return acc;
-              },
-              {}
-            ),
-          }))
-        );
-      }
-      data = transformAPIResponse[viewType](
-        payload.response,
-        aggregationLevel,
-        metricType
-      );
-      return {
-        data,
-        probeType: probeView,
-      };
-    });
-  },
-
-  updateStoreAfterDataIsReceived(data, probeType, appStore) {
-    // This function is called directly after the response has been received by
-    // the frontend. It will always run, even against cached data, as a way of
-    // resetting the necessary state.
-    const viewType = probeType;
-    let etc = {};
-
-    // filter out true/false aggregate results in boolean metrics. See: https://github.com/mozilla/glam/pull/1525#discussion_r694135079
-    if (data[0].metric_type === 'boolean') {
-      // eslint-disable-next-line no-param-reassign
-      data = data.filter((di) => di.client_agg_type === '');
-    }
-
-    if (viewType === 'categorical') {
-      etc = extractBucketMetadata(data);
-      appStore.setField('activeBuckets', etc.initialBuckets);
-    }
-
-    return { data, viewType, ...etc };
-  },
-  transformProbeForGLAM(probe) {
-    const pr = { ...probe };
-    return pr;
-  },
-  setDefaultsForProbe(store) {
-    const state = store.getState();
-    const { probe } = state;
-    if (!probe.send_in_pings.includes(state.productDimensions.ping_type)) {
-      // Try not to pick the 'All' option, at first.
-      const index = probe.send_in_pings.length > 1 ? 1 : 0;
-      let newPingtype = probe.send_in_pings[index];
-      if (
-        !Object.keys(this.dimensions.ping_type.values).includes(newPingtype)
-      ) {
-        // We don't know about this ping type, default to 'All'.
-        newPingtype = '*';
-      }
-      store.setDimension('ping_type', newPingtype);
-    }
-  },
-};
